third_party/WebKit/LayoutTests/fast/css/all-shorthand-css-text-expected.txt

Test for crbug.com/172051: all shorthand property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cssRules.length is 18
PASS cssRules[0].cssText is "div { all: initial; }"
PASS cssRules[1].cssText is not "div { all: initial; color: red; }"
PASS contains(cssRules[1].cssText, "color: red;") is true
PASS contains(cssRules[1].cssText, "background: initial;") is true
PASS contains(cssRules[1].cssText, "direction: ") is false
PASS contains(cssRules[1].cssText, "unicode-bidi: ") is false
PASS cssRules[2].cssText is "div { all: initial; }"
PASS cssRules[3].cssText is "div { all: initial !important; }"
PASS cssRules[4].cssText is not "div { all: initial; color: red !important; }"
PASS contains(cssRules[4].cssText, "color: red !important;") is true
PASS contains(cssRules[4].cssText, "background: initial;") is true
PASS contains(cssRules[4].cssText, "direction: ") is false
PASS contains(cssRules[4].cssText, "unicode-bidi: ") is false
PASS cssRules[5].cssText is "div { all: inherit; }"
PASS cssRules[6].cssText is not "div { all: inherit; color: red; }"
PASS contains(cssRules[6].cssText, "color: red;") is true
PASS contains(cssRules[6].cssText, "direction: ") is false
PASS contains(cssRules[6].cssText, "unicode-bidi: ") is false
PASS cssRules[7].cssText is "div { all: inherit; }"
PASS cssRules[8].cssText is "div { all: inherit !important; }"
PASS cssRules[9].cssText is not "div { all: inherit; color: red !important; }"
PASS contains(cssRules[9].cssText, "color: red !important") is true
PASS contains(cssRules[9].cssText, "background: inherit;") is true
PASS contains(cssRules[9].cssText, "direction: ") is false
PASS contains(cssRules[9].cssText, "unicode-bidi: ") is false
PASS cssRules[10].cssText is "div { }"
PASS cssRules[11].cssText is "div { all: initial; direction: rtl; }"
PASS cssRules[12].cssText is "div { direction: ltr; unicode-bidi: bidi-override; all: initial !important; }"
PASS contains(cssRules[13].cssText, "all: initial;") is false
PASS contains(cssRules[13].cssText, "-internal-") is false
PASS contains(cssRules[13].cssText, "direction: ltr;") is true
PASS contains(cssRules[13].cssText, "color: red;") is true
PASS contains(cssRules[14].cssText, "all: initial;") is false
PASS contains(cssRules[14].cssText, "-internal-") is false
PASS contains(cssRules[14].cssText, "font: initial;") is false
PASS contains(cssRules[14].cssText, "font-family: initial;") is true
PASS contains(cssRules[15].cssText, "all: initial;") is false
PASS contains(cssRules[15].cssText, "-internal-") is false
PASS contains(cssRules[14].cssText, "font: initial;") is false
PASS contains(cssRules[14].cssText, "marker: initial;") is true
PASS contains(cssRules[15].cssText, "width: inherit;") is true
PASS contains(cssRules[15].cssText, "min-width: initial;") is true
PASS contains(cssRules[15].cssText, "max-width: initial;") is true
PASS contains(cssRules[15].cssText, "marker-start: initial;") is false
PASS contains(cssRules[15].cssText, "marker-mid: initial;") is false
PASS contains(cssRules[15].cssText, "marker-end: initial;") is false
PASS cssRules[16].cssText is "div { all: unset; }"
PASS contains(cssRules[17].cssText, "all: unset;") is false
PASS contains(cssRules[17].cssText, "color: red;") is true
PASS contains(cssRules[17].cssText, "display: unset;") is true
PASS successfullyParsed is true

TEST COMPLETE
